requirements.txt
pandas
pyaudio
pygame
SpeechRecognition
openai
python-dotenv